Extract the 'do reboot' cmd from the DB. Renamed 'do reboot' to 'DoReload'.
