+++ src/cmd_god/godtest.cpp
@@ -22,4 +22,3 @@ void do_godtest(CharData *ch, char */*argument*/, int /* cmd */, int /* subcmd *
 }
 
 // vim: ts=4 sw=4 tw=0 noet syntax=cpp :
-
