+++ src/cmd_god/godtest.h
@@ -6,3 +6,5 @@ class CharData;
 void do_godtest(CharData *ch, char *argument, int /* cmd */, int /* subcmd */);
 
 #endif //BYLINS_GODTEST_H
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
