+++ src/cmd_god/reload.h
@@ -1,0 +1,13 @@
+//
+// Created by Sventovit on 04.09.2024.
+//
+
+#ifndef BYLINS_SRC_CMD_GOD_RELOAD_H_
+#define BYLINS_SRC_CMD_GOD_RELOAD_H_
+
+class CharData;
+void DoReload(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
+
+#endif //BYLINS_SRC_CMD_GOD_RELOAD_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
