+++ src/cmd_god/tabulate.h
@@ -9,3 +9,5 @@ class CharData;
 void DoTabulate(CharData *ch, char *argument, int/* cmd*/, int/* subcmd*/);
 
 #endif //BYLINS_SRC_CMD_GOD_TABULATE_H_
+
+// vim: ts=4 sw=4 tw=0 noet syntax=cpp :
